requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-apache-spark==4.9.0
pyspark==3.5.1
pytest==8.3.2
requests==2.32.3